# server/swagger/swagger.yaml
openapi: 3.0.0
info:
  description: |
    UPchieve backend API
    [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net, #swagger](http://swagger.io/irc/).
  version: "1.0.0"
  title: UPchieve
  contact:
    email: tech@upchieve.org
  license:
    name: Non-Profit Open Software License 3.0
    url: 'https://gitlab.com/upchieve/subway/-/blob/main/LICENSE'
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/upchieveorg/subway/1.0.0
  - url: 'https://app.upchieve.org/v1'
...
paths:
...
  /auth/login:
    post:
      tags:
        - authorization
      summary: Validates user email and password for login
      description: |
        Authenticates the user and grants a session if credentials are correct for login
      security:
        - cookieAuth: []
      requestBody:
        $ref: '#/components/requestBodies/EmailPassword'
      responses:
        '200':
          description: Authenticates the user with a session if credentials are correct and returns a user object
        '401':
          description: Unsuccessful authentication of user credentials
  /auth/logout:
    get:
      tags:
        - authorization
      summary: User logout
      description: |
        Removes the user's current session.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User is logged out and session is ended
          content:
            application/json:
              example:
                msg: "You have been logged out"
  /auth/register/checkcred:
    post:
      tags:
        - authorization
      summary: Validate user credentials for registration
      description: |
        Check whether the credential user entered is valid. (first step of registration) The server will check for email duplications and validate the password.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        $ref: '#/components/requestBodies/EmailPassword'
      responses:
        '200':
          description: Email is not already in use and password is validated
        '409':
          description: Email already exists
        '422':
          description: Email/password not provided or password is not valid
  /auth/register/student/open:
    post:
      tags:
        - authorization
      summary: Register new open student account
      description: |
        Creates a new account for a student from given parameters in request body
      security:
        - cookieAuth: []
      requestBody:
        description: Object is created after a form is correctly completed
        required: true
        content:
          application/json:
            schema:
                type: object
                properties:
                  email:
                    type: string
                  password:
                    type: string
                  highSchoolId:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  terms:
                    type: boolean
      responses:
        '200':
          description: Successful student registration and log in
        '403':
          description: |
            A 403 is generated when the backend identifies the request IP
            as coming from a non-US IP
        '422':
          description: |
            - False value for term agreement
            - Email or password are invalid
            - If an account with this email already exists
            - If your high school isn't approved
            - If any of the parameters are missing
        '500':
          description: Server error
  /auth/register/student/partner:
    post:
      tags:
        - authorization
      summary: Register new partner student account
      description: |
        Creates a new account for a student from given parameters in request body
      security:
        - cookieAuth: [ ]
      requestBody:
        description: Object is created after a form is correctly completed
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                studentPartnerOrg:
                  type: string
                partnerUserId:
                  type: string
                highSchoolId:
                  type: string
                partnerSite:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                college:
                  type: string
                terms:
                  type: boolean
              required:
                - email
                - password
                - firstName
                - lastName
                - terms
      responses:
        '200':
          description: Successful student registration and log in
        '422':
          description: |
            - False value for term agreement
            - Email or password are invalid
            - If an account with this email already exists
            - If any partner information isn't valid
            - If any of the parameters are missing
        '500':
          description: Server error
  /auth/register/volunteer/open:
    post:
      tags:
        - authorization
      summary: Register new open volunteer account
      description: |
        Creates an open volunteer account
      security:
        - cookieAuth: []
      requestBody:
        description: Object is created after a form is correctly completed
        required: true
        content:
          application/json:
            schema:
                type: object
                properties:
                  email:
                    type: string
                  password:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  phone:
                    type: string
                  terms:
                    type: boolean
      responses:
        '200':
          description: Successful volunteer registration and log in
        '422':
          description: |
            Any of the following can generate status code 422:
             - Missing value for term agreement
             - Missing email or password
             - Invalid password
  /auth/register/volunteer/partner:
    post:
      tags:
        - authorization
      summary: Register new partner volunteer account
      description: |
        Creates an partner volunteer account
      security:
        - cookieAuth: []
      requestBody:
        description: Object is created after a form is correctly completed
        required: true
        content:
          application/json:
            schema:
                type: object
                properties:
                  email:
                    type: string
                  password:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  phone:
                    type: string
                  terms:
                    type: boolean
                  volunteerPartnerOrg:
                    type: string
                  referredByCode:
                    type: string
      responses:
        '200':
          description: Successful volunteer registration and log in
        '422':
          description: |
            Any of the following can generate status code 422:
             - Missing value for term agreement
             - Missing email or password
             - Invalid password
             - Invalid volunteer Partner
             - Invalid partner domain
  /auth/partner/volunteer:
    get:
      tags:
        - authorization
      summary: Get partner manifest for volunteer sign-up
      description: |
        Confirms partner id from link and return partner manifest
      security:
        - cookieAuth: []
      parameters:
        - name: partnerId
          in: query
          description: Where `PARTNER_ID` is the key name of the volunteer partner organization defined in `config.ts` under `volunteerPartnerManifests`
          schema:
            type: string
            example: "example3"
      responses:
        '200':
          description: Returns a volunteer partner manifest object
        '404':
          description: No manifest found for volunteer partner id
  /auth/partner/student:
    get:
      tags:
        - authorization
      summary: Get partner manifest for student sign-up
      description: |
        Confirms partner id from link and return partner manifest
      security:
        - cookieAuth: []
      parameters:
        - name: partnerId
          in: query
          description: Where `PARTNER_ID` is the key name of the volunteer partner organization defined in `config.ts` under `volunteerPartnerManifests`
          schema:
            type: string
            example: "example3"
      responses:
        '200':
          description: Returns a student partner manifest object
        '404':
          description: No manifest found for student partner id
  /auth/reset/send:
    post:
      tags:
        - authorization
      summary: Requests a password reset email
      description: |
        Sends an email to verify the current user with unique hash. The email provided
        will overwrite the user record's email, in the event that the two do not match.
      security:
        - cookieAuth: []
      requestBody:
        description: Object is created after a form is correctly completed
        required: true
        content:
          application/json:
            schema:
                type: object
                properties:
                  email:
                    $ref: '#/components/schemas/Email'
      responses:
        '200':
          description: Request successful and password reset instructions sent
        '422':
          description: Missing email for password reset
  /auth/reset/confirm:
    post:
      tags:
        - authorization
      summary: Password verification
      description: |
        Confirms new password during reset
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
                type: object
                properties:
                  email:
                    $ref: '#/components/schemas/Email'
                  password:
                    $ref: "#/components/schemas/Email"
                  newpassword:
                    type: string
                  token:
                    type: string
      responses:
        '200':
          description: Token is accepted and password is changed
        '422':
          description: |
            The following will generate status code 422:
            - No token given in request body
            - Missing email or password
            - Must reenter password
            - Passwords do not match
...
  /auth/partner/student/code:
    get:
      tags:
        - authorization
      summary: Retrieves student partner code
      description: |
        Retrieves student partner code
        Where `PARTNER_SIGNUP_CODE` is equal to a `signupCode` defined in `config.ts` under `studentPartnerManifests`
      security:
        - cookieAuth: []
      parameters:
        - name: partnerSignupCode
          required: true
          in: query
          schema:
            type: string
            example: 'EX1'
      responses:
        '200':
          description: Returns object with studentPartnerKey
  /auth/partner/student-partners:
    get:
      tags:
        - authorization
      summary: Retrieves a list of student partners
      description: |
        - For admin only
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Returns object with a list of all student partners
  /auth/partner/volunteer-partners:
    get:
      tags:
        - authorization
      summary: Retrieves a list of volunteer partners
      description: |
        - For admin only
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Returns object with a list of all volunteer partners
...
components:
  schemas:
...
    Email:
      type: string
      example: "student1@upchieve.org"
...
    Password:
      type: string
      example: "Password123"
...
  requestBodies:
...
    EmailPassword:
      required: true
      content:
        application/json:
          schema:
              type: object
              properties:
                email:
                  $ref: '#/components/schemas/Email'
                password:
                  $ref: '#/components/schemas/Password'
...